Added check permissions before ddl operation

+++ src/main/java/api/services/AdminService.java
@@ -3,17 +3,22 @@ package api.services;
 import api.models.Course;
 import api.models.Group;
 import api.models.Subject;
+import api.models.User;
 import api.repositories.CourseRepository;
 import api.repositories.GroupRepository;
 import api.repositories.SubjectRepository;
 import api.repositories.UserRepository;
+import api.utils.error.PermissionDeniedException;
 import api.utils.info.*;
 import org.jetbrains.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
+
+import static api.controllers.SessionController.USER_ID;
 
 /**
  * Created by Vileven on 14.05.17.
@@ -37,8 +42,19 @@ public class AdminService {
         this.template = template;
     }
 
-    public Course createCourse(CourseInfo courseData) {
+    private boolean isAdmin(HttpSession session) {
+        final User user = userRepository.find((Long) session.getAttribute(USER_ID));
+        return user != null && user.getRole() == 0;
+
+    }
+
+    public Course createCourse(CourseInfo courseData, HttpSession session) throws PermissionDeniedException {
+        if (!isAdmin(session)) {
+            throw new PermissionDeniedException("permission denied");
+        }
+
         return courseRepository.create(new Course(courseData.getName()));
+
     }
 
     @Nullable
@@ -56,15 +72,26 @@ public class AdminService {
         return courseRepository.find(id);
     }
 
-    public Course updateCourse(CourseInfo courseData) {
+    public Course updateCourse(CourseInfo courseData, HttpSession session) throws PermissionDeniedException {
+        if (!isAdmin(session)) {
+            throw new PermissionDeniedException("permission denied");
+        }
+
         return courseRepository.update(new Course(courseData.getName()));
     }
 
-    public void deleteCourse(IdInfo idInfo) {
+    public void deleteCourse(IdInfo idInfo, HttpSession session) throws PermissionDeniedException {
+        if (!isAdmin(session)) {
+            throw new PermissionDeniedException("permission denied");
+        }
+
         courseRepository.delete(idInfo.getId());
     }
 
-    public List<Course> selectWithParams(SelectParametersInfo info) {
+    public List<Course> selectWithParams(SelectParametersInfo info, HttpSession session) throws PermissionDeniedException {
+        if (!isAdmin(session)) {
+            throw new PermissionDeniedException("permission denied");
+        }
         return courseRepository.selectWithParams(info.getLimit(), info.getOffset(), info.getOrders(), info.getFilters());
     }
 }
